Fix some matfree tests
We are now hitting an issue where we fail to handle empty kernels. It is
advantageous to know how many kernels there are *prior* to compiling
TSFC because this information is useful for generating the wrapper
kernel (which needs to relate form information to BC information). I
believe that the solution will be to hold onto empty kernels (`None`)
until later in the code generation pipeline.

In the meantime I have added assertions to catch this error.