Handle empty kernels from TSFC
We need to know how many kernels we are dealing with because the unroll
parameter for each kernel varies depending upon the boundary conditions.
This was previously not possible to handle since we would drop the empty
kernels inside of TSFC so we did not know how many we would be getting
back. To avoid needing an additional cache for TSFC kernels we now
return empty kernels so this can be handled at the right stage.