Fix extrusion bug
Previously I was using a too restrictive check for whether or not to
compute the offset. This led to some tests failing because the generated
code did not believe there to be one. We now determine whether or not to
use offset based upon the iterset.

In future the idea is to attach this metadata to the UFL so we don't
need to inspect Firedrake/PyOP2-level objects.