Fix some MG tests
The problem here is that we use COFFEE for code generation so we are
using the older interface for TSFC that previously did not have kernel
arguments configured. I now set this to `None` because we don't actually
need them since we use the legacy `par_loop` method in PyOP2.