Temporary fix for some extruded tests
So it turns out that we get cache hits when we shouldn't do depending
on whether or not we are iterating over a subset of an extruded set. Is
this information that should be encoded in the form or is this one of
those things that can be altered 'after the fact' like we do for
unrolled boundary conditions?